pull out jaxrs tracing to separate package
