+++ settings.gradle
@@ -13,3 +13,4 @@ include 'pkcs1-reader-sun'
 include 'retrofit2-clients'
 include 'tracing'
 include 'tracing-okhttp3'
+include 'tracing-jaxrs'
