+++ tracing/src/test/java/com/palantir/remoting3/tracing/TracersTest.java
@@ -21,17 +21,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.google.common.collect.Lists;
 import com.palantir.remoting.api.tracing.OpenSpan;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.StreamingOutput;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.MockitoAnnotations;
@@ -135,26 +130,6 @@ public final class TracersTest {
         });
         Tracer.startSpan("after-construction");
         callable.call();
-    }
-
-    @Test
-    public void testWrappingStreamingOutput_streamingOutputTraceIsIsolated() throws Exception {
-        Tracer.startSpan("outside");
-        StreamingOutput streamingOutput = Tracers.wrap((os) -> {
-            Tracer.startSpan("inside"); // never completed
-        });
-        streamingOutput.write(new ByteArrayOutputStream());
-        assertThat(Tracer.completeSpan().get().getOperation()).isEqualTo("outside");
-    }
-
-    @Test
-    public void testWrappingStreamingOutput_traceStateIsCapturedAtConstructionTime() throws Exception {
-        Tracer.startSpan("before-construction");
-        StreamingOutput streamingOutput = Tracers.wrap((os) -> {
-            assertThat(Tracer.completeSpan().get().getOperation()).isEqualTo("before-construction");
-        });
-        Tracer.startSpan("after-construction");
-        streamingOutput.write(new ByteArrayOutputStream());
     }
 
     @Test
@@ -270,20 +245,6 @@ public final class TracersTest {
         return new Runnable() {
             @Override
             public void run() {
-                assertThat(Tracer.getTraceId()).isEqualTo(expectedTraceId);
-                assertThat(getCurrentFullTrace()).isEqualTo(expectedTrace);
-                assertThat(MDC.get(Tracers.TRACE_ID_KEY)).isEqualTo(expectedTraceId);
-            }
-        };
-    }
-
-    private static StreamingOutput traceExpectingStreamingOutput() {
-        final String expectedTraceId = Tracer.getTraceId();
-        final List<OpenSpan> expectedTrace = getCurrentFullTrace();
-
-        return new StreamingOutput() {
-            @Override
-            public void write(OutputStream output) throws IOException, WebApplicationException {
                 assertThat(Tracer.getTraceId()).isEqualTo(expectedTraceId);
                 assertThat(getCurrentFullTrace()).isEqualTo(expectedTrace);
                 assertThat(MDC.get(Tracers.TRACE_ID_KEY)).isEqualTo(expectedTraceId);
